backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
authlib
python-jose
python-multipart
bcrypt
passlib
httpx
